Fix AttributeError: 'Namespace' object has no attribute 'base_dir'
Traceback (most recent call last):
  File "/MultiC/multic/cli/MultiCompartmentSegment/MultiCompartmentSegment.py", line 71, in <module>
    main(CLIArgumentParser().parse_args())
  File "/MultiC/multic/cli/MultiCompartmentSegment/MultiCompartmentSegment.py", line 68, in main
    run_it(args)
  File "../segmentationschool/segmentation_school.py", line 45, in run_it
    main(args)
  File "../segmentationschool/segmentation_school.py", line 37, in main
    predict(args=args)
  File "../segmentationschool/Codes/IterativePredict_1X.py", line 198, in predict
    if args.base_dir and os.path.exists(args.base_dir):
AttributeError: 'Namespace' object has no attribute 'base_dir'
DockerException: Non-zero exit code from docker container (1).
  File "/opt/worker_venv/lib/python3.10/site-packages/celery/app/trace.py", line 453, in trace_task
    R = retval = fun(*args, **kwargs)
  File "/opt/worker_venv/lib/python3.10/site-packages/slicer_cli_web/girder_worker_plugin/direct_docker_run.py", line 107, in __call__
    super().__call__(*args, **kwargs)
  File "/opt/worker_venv/lib/python3.10/site-packages/girder_worker/docker/tasks/__init__.py", line 345, in __call__
    super().__call__(*args, **kwargs)
  File "/opt/worker_venv/lib/python3.10/site-packages/girder_worker/task.py", line 155, in __call__
    results = super().__call__(*_t_args, **_t_kwargs)
  File "/opt/worker_venv/lib/python3.10/site-packages/celery/app/trace.py", line 736, in __protected_call__
    return self.run(*args, **kwargs)
  File "/opt/worker_venv/lib/python3.10/site-packages/slicer_cli_web/girder_worker_plugin/direct_docker_run.py", line 142, in run
    return _docker_run(task, **kwargs)
  File "/opt/worker_venv/lib/python3.10/site-packages/girder_worker/docker/tasks/__init__.py", line 437, in _docker_run
    _run_select_loop(task, container, read_streams, write_streams)
  File "/opt/worker_venv/lib/python3.10/site-packages/girder_worker/docker/tasks/__init__.py", line 258, in _run_select_loop
    raise DockerException('Non-zero exit code from docker container (%d).' % exit_code)

+++ multic/segmentationschool/Codes/IterativePredict_1X.py
@@ -195,7 +195,7 @@ def predict(args):
 
         slide.close()
 
-        if args.base_dir and os.path.exists(args.base_dir):
+        if "base_dir" in args and args.base_dir and os.path.exists(args.base_dir):
             mask_filename = args.base_dir + "/" + dirs['fileID'] + ".png"
             print(f"Writing mask to file: {mask_filename}")
             cv2.imwrite(mask_filename, wsiMask)
